+++ examples/rebuild_shared_projection/src/main.rs
@@ -1,13 +1,13 @@
-use aggregate_merging::aggregates::{AggregateA, AggregateB};
-use aggregate_merging::projectors::{Counter, CounterProjector};
-use aggregate_merging::structs::{CommandA, CommandB, CounterError, ProjectorEvent};
-use esrs::aggregate::{AggregateManager, AggregateState};
-use esrs::projector::PgProjector;
-
 use futures::StreamExt;
 use sqlx::migrate::MigrateDatabase;
-use sqlx::{pool::PoolOptions, Pool, Postgres};
+use sqlx::{pool::PoolOptions, Pool, Postgres, Transaction};
 use uuid::Uuid;
+
+use aggregate_merging::aggregates::{AggregateA, AggregateB};
+use aggregate_merging::projectors::Counter;
+use aggregate_merging::structs::{CommandA, CommandB, CounterError, EventA, EventB};
+use esrs::aggregate::{AggregateManager, AggregateState};
+use esrs::store::StoreEvent;
 
 // A simple example demonstrating rebuilding a read-side projection from an event
 // stream
@@ -25,69 +25,64 @@ async fn main() {
     let count_id = Uuid::new_v4();
 
     setup(&pool, count_id).await;
-    let agg_a = AggregateA::new(&pool).await.unwrap();
-    let agg_b = AggregateB::new(&pool).await.unwrap();
-    let store_a = agg_a.event_store();
-    let store_b = agg_b.event_store();
 
-    let mut events_a = store_a.get_all().map(|r| match r {
-        Ok(e) => Ok(e.map(ProjectorEvent::from)),
-        Err(e) => Err(e),
-    });
-    let mut events_b = store_b.get_all().map(|r| match r {
-        Ok(e) => Ok(e.map(ProjectorEvent::from)),
-        Err(e) => Err(e),
-    });
+    let store_a = AggregateA::new_store(&pool).await.unwrap();
+    let store_b = AggregateB::new_store(&pool).await.unwrap();
 
-    let projectors: Vec<Box<dyn PgProjector<ProjectorEvent, CounterError>>> = vec![Box::new(CounterProjector)];
+    let mut events_a = store_a.get_all();
+    let mut events_b = store_b.get_all();
 
-    let mut a = None;
-    let mut b = None;
+    let mut event_a_opt: Option<Result<StoreEvent<EventA>, CounterError>> = events_a.next().await;
+    let mut event_b_opt: Option<Result<StoreEvent<EventB>, CounterError>> = events_b.next().await;
+
+    let mut transaction: Transaction<Postgres> = pool.begin().await.expect("Failed to create transaction");
+
     loop {
-        if a.is_none() {
-            a = events_a.next().await;
-        }
-        if b.is_none() {
-            b = events_b.next().await;
-        }
-        if a.is_none() && b.is_none() {
-            break;
-        }
-        let mut transcation = pool.begin().await.unwrap();
-        for projector in &projectors {
-            if a.is_none() {
-                projector
-                    .project(b.as_ref().unwrap().as_ref().unwrap(), &mut transcation)
-                    .await
-                    .unwrap();
-                b = None;
-                continue;
+        let a_opt: Option<&StoreEvent<EventA>> = event_a_opt.as_ref().map(|v| v.as_ref().unwrap());
+        let b_opt: Option<&StoreEvent<EventB>> = event_b_opt.as_ref().map(|v| v.as_ref().unwrap());
+
+        match (a_opt, b_opt) {
+            (Some(a), Some(b)) if a.occurred_on <= b.occurred_on => {
+                for projector in store_a.projectors() {
+                    projector
+                        .project(a, &mut *transaction)
+                        .await
+                        .expect("Projector A error");
+                }
+
+                event_a_opt = events_a.next().await;
             }
-            if b.is_none() {
-                projector
-                    .project(a.as_ref().unwrap().as_ref().unwrap(), &mut transcation)
-                    .await
-                    .unwrap();
-                a = None;
-                continue;
+            (Some(a), None) => {
+                for projector in store_a.projectors() {
+                    projector
+                        .project(a, &mut *transaction)
+                        .await
+                        .expect("Projector A error");
+                }
+
+                event_a_opt = events_a.next().await;
             }
-            let a_inner = a.as_ref().unwrap().as_ref().unwrap();
-            let b_inner = b.as_ref().unwrap().as_ref().unwrap();
-            if a_inner.occurred_on > b_inner.occurred_on {
-                projector.project(b_inner, &mut transcation).await.unwrap();
-                b = None;
-            } else {
-                projector.project(a_inner, &mut transcation).await.unwrap();
-                a = None;
+            (Some(_), Some(b)) | (None, Some(b)) => {
+                for projector in store_b.projectors() {
+                    projector
+                        .project(b, &mut *transaction)
+                        .await
+                        .expect("Projector B error");
+                }
+
+                event_b_opt = events_b.next().await;
             }
-        }
-        transcation.commit().await.unwrap();
+            (None, None) => break,
+        };
     }
+
+    transaction.commit().await.unwrap();
 
     let counter = Counter::by_id(count_id, &pool)
         .await
         .expect("Failed to retrieve counter")
         .expect("Failed to find counter");
+
     assert!(counter.count_a == 1 && counter.count_b == 1);
 }
