+++ src/esrs/postgres/mod.rs
@@ -3,6 +3,8 @@ use std::future::Future;
 
 use async_trait::async_trait;
 use chrono::{DateTime, Utc};
+use futures::stream::BoxStream;
+use futures::StreamExt;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use sqlx::postgres::PgQueryResult;
@@ -11,12 +13,14 @@ use sqlx::{Executor, Pool, Postgres, Transaction};
 use uuid::Uuid;
 
 use crate::aggregate::Aggregate;
+use crate::esrs::postgres::statement::Statements;
 use crate::esrs::store::{EventStore, StoreEvent};
 use crate::esrs::{event, SequenceNumber};
 
 pub mod policy;
 pub mod projector;
 
+mod statement;
 #[cfg(test)]
 mod tests;
 
@@ -24,8 +28,8 @@ type Projector<Event, Error> = Box<dyn projector::Projector<Event, Error> + Send
 type Policy<Event, Error> = Box<dyn policy::Policy<Event, Error> + Send + Sync>;
 
 pub struct PgStore<Event, Error> {
-    table_name: String,
     pool: Pool<Postgres>,
+    statements: Statements,
     projectors: Vec<Projector<Event, Error>>,
     policies: Vec<Policy<Event, Error>>,
 }
@@ -42,8 +46,8 @@ where
         policies: Vec<Policy<Event, Error>>,
     ) -> Self {
         Self {
-            table_name: format!("{}_events", A::name()),
             pool: pool.clone(),
+            statements: Statements::new(A::name()),
             projectors,
             policies,
         }
@@ -51,7 +55,16 @@ where
 
     // TODO: doc
     pub async fn setup(self) -> Result<Self, Error> {
-        create_table::<Error>(&self.table_name, &self.pool).await?;
+        // Create events table if not exists
+        let _: PgQueryResult = sqlx::query(self.statements.create_table()).execute(&self.pool).await?;
+        // Create index on aggregate_id for `by_aggregate_id` query.
+        let _: PgQueryResult = sqlx::query(self.statements.create_index()).execute(&self.pool).await?;
+
+        // Create unique constraint `aggregate_id`-`sequence_number` to avoid race conditions.
+        let _: PgQueryResult = sqlx::query(self.statements.create_unique_constraint())
+            .execute(&self.pool)
+            .await?;
+
         Ok(self)
     }
 
@@ -66,12 +79,7 @@ where
     ) -> Result<StoreEvent<Event>, Error> {
         let id: Uuid = Uuid::new_v4();
 
-        let insert_query: String = format!(
-            "INSERT INTO {} (id, aggregate_id, payload, occurred_on, sequence_number) VALUES ($1, $2, $3, $4, $5)",
-            &self.table_name
-        );
-
-        let _ = sqlx::query(insert_query.as_str())
+        let _ = sqlx::query(self.statements.insert())
             .bind(id)
             .bind(aggregate_id)
             .bind(Json(&event))
@@ -86,6 +94,18 @@ where
             payload: event,
             occurred_on,
             sequence_number,
+        })
+    }
+
+    // TODO: doc
+    pub fn get_all<'a>(&self) -> BoxStream<Result<StoreEvent<Event>, Error>> {
+        Box::pin({
+            sqlx::query_as::<_, event::Event>(self.statements.select_all())
+                .fetch(&self.pool)
+                .map(|res| match res {
+                    Ok(event) => event.try_into().map_err(Into::into),
+                    Err(error) => Err(error.into()),
+                })
         })
     }
 
@@ -121,12 +141,7 @@ where
     Error: From<sqlx::Error> + From<serde_json::Error> + Send + Sync,
 {
     async fn by_aggregate_id(&self, aggregate_id: Uuid) -> Result<Vec<StoreEvent<Event>>, Error> {
-        let by_aggregate_id_query: String = format!(
-            "SELECT * FROM {} WHERE aggregate_id = $1 ORDER BY sequence_number ASC",
-            &self.table_name
-        );
-
-        Ok(sqlx::query_as::<_, event::Event>(by_aggregate_id_query.as_str())
+        Ok(sqlx::query_as::<_, event::Event>(self.statements.by_aggregate_id())
             .bind(aggregate_id)
             .fetch_all(&self.pool)
             .await?
@@ -178,9 +193,7 @@ where
     async fn delete_by_aggregate_id(&self, aggregate_id: Uuid) -> Result<(), Error> {
         let mut transaction: Transaction<Postgres> = self.pool.begin().await?;
 
-        let delete_by_aggregate_id_query: String = format!("DELETE FROM {} WHERE aggregate_id = $1", &self.table_name);
-
-        let _ = sqlx::query(delete_by_aggregate_id_query.as_str())
+        let _ = sqlx::query(self.statements.delete_by_aggregate_id())
             .bind(aggregate_id)
             .execute(&mut *transaction)
             .await
@@ -194,66 +207,4 @@ where
 
         Ok(())
     }
-
-    fn get_all(&self) -> BoxStream<Result<StoreEvent<Event>, Error>> {
-        Box::pin(
-            sqlx::query_as::<_, event::Event>(self.queries.select_all())
-                .fetch(&self.pool)
-                .map(|res| match res {
-                    Ok(e) => match e.try_into() {
-                        Ok(e) => Ok(e),
-                        Err(e) => Err(Error::from(e)),
-                    },
-                    Err(e) => Err(e.into()),
-                }),
-        )
-    }
-}
-
-async fn create_table<Error>(table_name: &str, pool: &Pool<Postgres>) -> Result<(), Error>
-where
-    Error: From<sqlx::Error> + Send + Sync,
-{
-    // Create events table if not exists
-    let create_table_query: String = format!(
-        "CREATE TABLE IF NOT EXISTS {0}
-            (
-              id uuid NOT NULL,
-              aggregate_id uuid NOT NULL,
-              payload jsonb NOT NULL,
-              occurred_on TIMESTAMPTZ NOT NULL DEFAULT current_timestamp,
-              sequence_number INT NOT NULL DEFAULT 1,
-              CONSTRAINT {0}_pkey PRIMARY KEY (id)
-            )",
-        table_name
-    );
-
-    let _: PgQueryResult = sqlx::query(create_table_query.as_str())
-        .bind(table_name)
-        .execute(pool)
-        .await?;
-
-    // Create index on aggregate_id for `by_aggregate_id` query.
-    let create_aggregate_id_index_query: String = format!(
-        "CREATE INDEX IF NOT EXISTS {0}_aggregate_id ON {0} USING btree (((payload ->> 'id'::text)))",
-        table_name
-    );
-
-    let _: PgQueryResult = sqlx::query(create_aggregate_id_index_query.as_str())
-        .bind(table_name)
-        .execute(pool)
-        .await?;
-
-    // Create unique constraint `aggregate_id`-`sequence_number` to avoid race conditions.
-    let create_aggregate_id_sequence_number_unique_constraint_query: String = format!(
-        "CREATE UNIQUE INDEX IF NOT EXISTS {0}_aggregate_id_sequence_number ON {0}(aggregate_id, sequence_number)",
-        table_name
-    );
-
-    let _: PgQueryResult = sqlx::query(create_aggregate_id_sequence_number_unique_constraint_query.as_str())
-        .bind(table_name)
-        .execute(pool)
-        .await?;
-
-    Ok(())
 }
